+++ tests/testthat/test-wsj_pal.R
@@ -14,7 +14,7 @@ test_that("theme_wsj works", {
 })
 
 test_that("theme_wsj raises error with invalid palette", {
-  expect_error(wsj_pal("asdgasa"), regexp = "must be one of")
+  expect_snapshot(wsj_pal("asdgasa"), error = TRUE)
 })
 
 test_that("scale_colour_wsj works", {